Fix parser for win

+++ Parser_main_win_cpp/CaffAnimation.cpp
@@ -10,17 +10,17 @@
 
 class CaffAnimation {
   private:
-    double duration;
+    long long duration;
     
   public:
     CaffAnimation() {
     }
     
-    void setDuration(double dur) {
+    void setDuration(long long dur) {
         duration = dur;
     }
     
-    long getDuration() {
+    long long getDuration() {
         return duration;
     }
